refactor(auth): rename auth web to a top-level open command
`oo auth web` read like a web-based way to log the CLI in, the direction `gh auth login --web` established, while this flow is the reverse: the CLI hands its session to the browser. `oo open` follows the `stripe open` and `heroku open` convention, where the direction is unambiguous. The behavior is unchanged, the command prints the URL and never launches a browser.

The shared `readCommandTelemetryProperties` test helper moved to _ __tests__/helpers.ts _ now that two CLI test files need it.
