+++ src/application/commands/catalog.ts
@@ -13,6 +13,7 @@ import { llmCommand } from "./llm/index.ts";
 import { logCommand } from "./log/index.ts";
 import { loginCommand } from "./login.ts";
 import { logoutCommand } from "./logout.ts";
+import { openCommand } from "./open.ts";
 import { searchCommand } from "./search.ts";
 import { skillsCommand } from "./skills/index.ts";
 import { teamCommand } from "./team/index.ts";
@@ -53,6 +54,7 @@ export function createCliCatalog(): CliCatalog {
             llmCommand,
             loginCommand,
             logoutCommand,
+            openCommand,
             teamCommand,
             completionCommand,
             configCommand,
